// ui/src/lib/api/generated/model/tag.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Centoire API
 * Centoire community platform API
 * OpenAPI spec version: 0.1.0
 */
import type { TagCategory } from './tagCategory';

export interface Tag {
  id: string;
  name: string;
  slug: string;
  category: TagCategory;
  /** @nullable */
  description: string | null;
  postCount: number;
  followerCount: number;
}